pyproject.toml
[project]
name = "LogosAI"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-genai>=1.19.0",
    "gradio[mcp]>=5.33.0",
    "langchain[google-genai]>=0.3.27",
    "langgraph>=0.6.4",
    "mistralai>=1.8.1",
    "openai>=1.84.0",
    "sqlalchemy>=2.0.42",
    "uvicorn[standard]>=0.34.3",
]
